VisiCalc orientado a objetos
