+++ entregas/andriuwYepez/reto-002/Main.java
@@ -1,0 +1,5 @@
+public class Main {
+    public static void main(String[] args) {
+        VisiCalcApp.main(args);
+    }
+}
